ds_prefill :: LB hang fix avoid typecast (#45716)
Addresses https://github.com/tenstorrent/tt-metal/issues/44928
Related to https://github.com/tenstorrent/tt-metal/issues/45714 and
https://github.com/tenstorrent/tt-metal/issues/41163
                                                                
- Remove ttnn.typecast(uint16 → int32) from the MoE path, which caused
data corruption under Tracy profiler
- MoE ops (dispatch, post_combine_reduce) now consume uint16 indices
natively from the gate, eliminating the buggy typecast
- Add `read_tile_value_uint16` compute kernel helper for reading uint16
elements from cb since the `read_tile_value` by default reads int32

Tested locally 200 iterations of layer3 block loop test ISL=3840, no
hang. ISL=3840 used to always hang on the 1st iteration.

### CI Status
_Auto-generated on every push. Badges update live. Click a badge to
filter runs by this branch._

-
[![](https://github.com/tenstorrent/tt-metal/actions/workflows/t3000-e2e-tests.yaml/badge.svg?branch=nmilicevic/deepseek-hang-fix)](https://github.com/tenstorrent/tt-metal/actions/workflows/t3000-e2e-tests.yaml?query=branch:nmilicevic/deepseek-hang-fix)
-
[![](https://github.com/tenstorrent/tt-metal/actions/workflows/blackhole-e2e-tests.yaml/badge.svg?branch=nmilicevic/deepseek-hang-fix)](https://github.com/tenstorrent/tt-metal/actions/workflows/blackhole-e2e-tests.yaml?query=branch:nmilicevic/deepseek-hang-fix)
[Perf tests no
hang](https://github.com/tenstorrent/tt-metal/actions/runs/26749717258/job/78953362078)
-
[![](https://github.com/tenstorrent/tt-metal/actions/workflows/demo-sp-release.yaml/badge.svg?branch=nmilicevic/deepseek-hang-fix)](https://github.com/tenstorrent/tt-metal/actions/workflows/demo-sp-release.yaml?query=branch:nmilicevic/deepseek-hang-fix)
[Perf tests no
hang](https://github.com/tenstorrent/tt-metal/actions/runs/26812841467/job/79051030131)

GPT OSS
-
[![](https://github.com/tenstorrent/tt-metal/actions/workflows/models-t1-e2e-tests.yaml/badge.svg?branch=nmilicevic/deepseek-hang-fix)](https://github.com/tenstorrent/tt-metal/actions/workflows/models-t1-e2e-tests.yaml?query=branch:nmilicevic/deepseek-hang-fix)

+++ models/demos/deepseek_v3_d_p/tests/pcc/test_reduce.py
@@ -163,7 +163,7 @@ def test_ttnn_reduce(
         mesh_mapper=indices_mapper,
         layout=ttnn.ROW_MAJOR_LAYOUT,
         device=mesh_device,
-        dtype=ttnn.int32,
+        dtype=ttnn.uint16,
     )
     tt_expert_dispatch_table = ttnn.from_torch(
         expert_dispatch_table,
